fix: getNexus function
