+++ src/sdk/modules/utils/Helpers.ts
@@ -10,10 +10,16 @@ import {
   pad,
   publicActions,
   toFunctionSelector,
-  toHex
+  toHex,
 } from "viem"
-import { ERROR_MESSAGES } from "../../account/index.js"
+import {
+  AddressConfig,
+  ERROR_MESSAGES,
+  NexusVersion,
+  semverCompare,
+} from "../../account/index.js"
 import type { AnyData, ModularSmartAccount } from "./Types.js"
+import { DEFAULT_CONFIGURATIONS_BY_NEXUS_VERSION } from "../../constants/index.js"
 
 /**
  * Represents a hardcoded hex value reference.
@@ -116,7 +122,7 @@ export const isPermitSupported = async (
       return client
         .call({
           to: tokenAddress,
-          data: `${selector}${padding}` as Hex
+          data: `${selector}${padding}` as Hex,
         })
         .then(() => true)
         .catch((error) => {
@@ -138,7 +144,7 @@ export const isPermitSupported = async (
       checkPermitEnabled(
         noncesSelector,
         `000000000000000000000000${"0".repeat(40)}`
-      )
+      ),
     ])
 
     return hasPermit && hasDomainSeparator && hasNonces
@@ -146,4 +152,30 @@ export const isPermitSupported = async (
     console.error("Error checking permit support:", err)
     return false
   }
+}
+
+/**
+ * Returns the appropriate configuration based on the SDK version
+ * @param version - The SDK version string (e.g., "0.2.0")
+ * @returns The configuration containing attester and factory addresses
+ * @throws Error if the version is not supported
+ */
+export function getNexus(nexusVersion: NexusVersion): AddressConfig {
+  // If the version is explicitly provided in the DEFAULT_CONFIGURATIONS_BY_VERSION mapping
+  if (nexusVersion in DEFAULT_CONFIGURATIONS_BY_NEXUS_VERSION) {
+    return DEFAULT_CONFIGURATIONS_BY_NEXUS_VERSION[nexusVersion]
+  }
+
+  // If the version is not explicitly listed, find the closest compatible version
+  // Sort the available versions in descending order
+  const allVersions = Object.keys(DEFAULT_CONFIGURATIONS_BY_NEXUS_VERSION).sort(
+    (a, b) => semverCompare(b, a)
+  )
+
+  // If no compatible version is found, throw an error
+  throw new Error(
+    `Unsupported Nexus version: ${nexusVersion}. Compatible versions are: ${allVersions.join(
+      ", "
+    )}`
+  )
 }
